move kernel debug parser to kernel crate
