+++ src/kernel/src/kernel_debug_parser.rs
