+++ src/util/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod array;
 pub mod io;
-pub mod kernel_debug_parser;
 pub mod macros;
 pub mod mem;
 pub mod slice;
